chore: update Next.js to version 16.2.1 and adjust TypeScript configuration

+++ app/api/units/[userId]/route.ts
@@ -7,10 +7,10 @@ import { isAdmin } from "@/lib/admin";
 
 export const GET = async (req: Request,
     { params }: {
-        params: Promise<{ unitsId: number }>
+        params: Promise<{ userId: string }>
     },
 ) => {
-    const { unitsId } = await params;
+    const { userId } = await params;
     const adminStatus = await isAdmin();
 
     if (!adminStatus) {
@@ -18,7 +18,7 @@ export const GET = async (req: Request,
     }
 
     const data = await db.query.units.findFirst({
-        where: eq(units.id, unitsId),
+        where: eq(units.id, Number(userId)),
     });
 
     return NextResponse.json(data);
@@ -26,10 +26,10 @@ export const GET = async (req: Request,
 
 export const PUT = async (req: Request,
     { params }: {
-        params: Promise<{ unitsId: number }>
+        params: Promise<{ userId: string }>
     },
 ) => {
-    const { unitsId } = await params;
+    const { userId } = await params;
     const adminStatus = await isAdmin();
 
     if (!adminStatus) {
@@ -39,17 +39,17 @@ export const PUT = async (req: Request,
     const body = await req.json();
     const data = await db.update(units).set({
         ...body,
-    }).where(eq(units.id, unitsId)).returning();
+    }).where(eq(units.id, Number(userId))).returning();
 
     return NextResponse.json(data[0]);
 };
 
 export const DELETE = async (req: Request,
     { params }: {
-        params: Promise<{ unitsId: number }>
+        params: Promise<{ userId: string }>
     },
 ) => {
-    const { unitsId } = await params;
+    const { userId } = await params;
     const adminStatus = await isAdmin();
 
     if (!adminStatus) {
@@ -57,7 +57,7 @@ export const DELETE = async (req: Request,
     }
 
     const data = await db.delete(units)
-        .where(eq(units.id, unitsId)).returning();
+        .where(eq(units.id, Number(userId))).returning();
 
     return NextResponse.json(data[0]);
 };
